handle missing dns-packet and user-flags
For reasons not yet known, userBlocklistFlagUint may go missing, and that
causes issues in dns-operation which don't at all expect it (null, undefined).

Fix that by treating empty, null, undefined strings as just empty, and
returning early from functions that can't do much with missing user-flags.

In a similar issue seen in prod, current-request doesn't do well with a missing
(undefined, null) decodedDnsPacket obj. So, before using it in current-request
make sure it is set to an equivalent empty obj.