avoid eager closing sockets & malformed servfails
In server.js, it isn't necessairly a good idea to close sockets if the request
buffer couldn't be flushed in its entirety to kernel. That is, treating
'false' socket.write response as a socket destorying thing can be avoided.
Destorying sockets now only on error events or internal exceptions. On invalid
however, requests, the socket is left open to give query-pipelining a chance.

On internal exceptions, current-request constructs a servfail response, which
was malformed because servfail packet was set in options and not in the body of
the response obj...