Event bus to streamline process up
Since nodejs doesn't respect import sequence (unlike deno), conditionals
are required to initialize global deps like log, env, plugin.services etc.

This is solved with a pub-sub mechanism where-in modules subscribe to events to
know when their deps are up and running. Today, the "config" module, imported
for side-effects from the entrypoints (server-workers.js for workers,
server-deno.ts for deno, and server-node.js for node) of respective runtimes,
initializes "log" and "env" objs, and kicks-off other basic setup. It then
publishes a "ready" event which triggers the "plugin" module, that then inits
essential "services". Once the service init is done, "plugin" publishes a "go"
event. The entrypoints listening on the "go" event now bring up their endpoints
as required and start serving DoH/DoT requests.

This could have been a state-machine, but it is impl as pub-sub, for now.