Async pub/sub for Workers
On Workers, the event listener for the main "fetch event" fires before the
microtask-queue responsible for broadcasting published events to subscribers
has had the chance to propogate. This results in 5xx responses, because the
state-of-the-system, modified by these subscribers, is not up-to-date.

Add a new broadcast mechanism which returns a promise that callers can then
await on, as an alternative to queuing callbacks on the microtask-queue. These
promises reject after a caller-supplied timeout elapses, or resolve whenever
its corresponding event fires.