timeout query with a serv-fail answer or a 4xx
Servfail answers are constructed as approp on errors with query-id & questions,
where possible. If not, the clients are sent a 4xx response instead of a 5xx:
Chromium-based and Firefox-based browser do not retry 5xx doh answers.

Response timeout timers were not closed once a dns query had been responded to
within specified timeout, but that is now dealt by a "deferred promise" which
resolves either on a timeout or on a valid response.