Fix servfail and handle zero-answer dns responses
DoH clients may send 0 IDs. But servfail refused to build a 0 ID responses.
This resulted in 4xx whenever servfails (which are 2xx) needed to be sent.
Besides, responses lacked content-length headers, too. This commit fixes these
issues in currentRequest.js.

In case a query resulted in zero-answers, the resolver incorrectly sent an
'exception' response. Incorrect because nxdomains, for example, have zero ans
but are valid responses to client for the given query nevertheless. The plugin
now simply forwards whatever the response was from upstream, instead of double
checking for the presence of answers (which may not be there).