Stamps from caches must be used as-is
Stamps were filtered out for domains present in question or answer, but that
is not only unnecessary but also breaks wildcard blocks that need subdomains
in addition to domains present in the dns question / answer (but filtering
did not account for this).

So: Simply returning the stamps stored in the cache metadata should not only
suffice, but it is infact the correct and valid behaviour. For the cases where
stamps aren't available in the cache, blocklist-filter is used to fetch them
for all subdomains of whatever domains are in the dns question / answer and
then stored in the caches.