Static, stable cache-key generation
Previously, the incoming request-url formed the cache-key, which resulted in
different cache-entries for the same dns query. Why? A request-url for DoT is
different to DoH, while DoH itself could be serving multiple domains / URLs
(for ex, subdomains of workers.dev, rethinkdns.com, and bravedns.com front
DoH on Cloudflare). This meant a ploriferation of entries even for the same
dns question (ie, key is https://a.rethinkdns.com/example.com if the incoming
request-url is a.rethinkdns.com and the dns query is example.com; whereas key
is https://b.bravedns.com/example.com for request-url b.bravedns.com for the
same query, example.com; and so on...)

Normalizing cache-key to a static-url like https://caches.rethinkdns.com/
removes dependency on the incoming request-url, and improves cache-hits.