Make util.emptyObj handle objs with inherited keys
util.emptyObj did not work for 'new URL(str)', 'new Date()', 'new Response()'
because even if these objects did not contain own properties, they did inherit
from their prototypes (something which the util.emptyObj impl did not account
for). This resulted in all sorts of wrong empty object detection, that is, an
obj (with prototypal properties) would be marked empty, even when it was not.