Share blockstamps from caches with other plugins
domain-name-cache which stored blockstamps against domain-names was removed
in a previous commit. This means, dns-resolver has to always traverse
the trie to get blockstamps if the caches, which already may have these stamps,
fail to pass those on via plugin params.

This change passes on stamps, if any, from the caches on to the dns-resolver.