Use http-cache only when blocklists not setup
http-cache api has 5ms to 10ms latency on average, and that's particularly
expensive on cache-miss, since upstream doh themselves average 5ms to 10ms
(and almost always have an answer for the query, unlike cache, which may not).

Since caches were introduced as a way to help avoid downloading and setting up
blocklists (an operation that takes anywhere between 100ms to 5s), perhaps it
is prudent to only lookup local-cache when they cost to set these blocklists
have already been paid by the process.

Note that, on Deno and Fly deploys, http-cache is unavailble anyway.