Blockstamp output flags incorrect
Two issues: The first one was that the output body needed to be reversed before
being converted into a uint16array. This was a regression from before @4dbdcd0a

The second one was that the incoming domain blockstamp was an array, and not
a uint16array as required by the method comparing and intersecting the two
blockstamps, the domain's and the user's.