Normalize return type of dnsResolver.js:resolveDns
Make resolveDns return Promise of an array of Promises consistently, so as to
avoid unhandled exceptions. The tricky thing is async function wrap any return
value in a Promise{} and so, a double 'await' is required to get to the array
of Promises it wraps. Another tricky thing is the unwrapping of Promises is
when a Promise{} wraps just a single other Promise{}, both unwrap on a single
await. But a Promise{Array<Promise>} doesn't unwrap on a single await... This
makes it important to consistently return an Array<Promise> for the caller to
then await twice and do so without any surprises.