blocklist-wrapper as a dep not a plugin
blocklist-wrapper is a plugin, and so it gets initialized like one, processed
like one, which is sequentially (as all plugins are) one after the other.

This prevents an opportunity for an optimization where the blocklist files
downloaded by blocklist-wrapper can be done in parallel with upstreaming dns
requests.
Blocklist files take anywhere between 100ms (from cloudflare caches) to 2s
(cloudfront caches) to 5s (s3); while upstreaming doh requests take 10ms to
30ms+. On Workers, up to 6 network-requests can be done in-parallel. But
because blocklist-wrapper is a plugin that runs before dns-operation,
these are ran sequentially (but offer no benefit).

Turning blocklist-wrapper into a dependency of dns-operation makes it easier
to parllalelize these network-requests.