Move to Module Workers
Workers support "module" workers, which are required to be able to use
DurableObjects. In module workers, env bindings are not global anymore but
instead get passed as an arg to the fetch event handler. As a result, the
current env setup that relies on bindings being global needs a change. As part
of the changes, 'env' global binding is removed in favour of envManager which
exposes 'get' method that retrieves runtime-specific values where present
or default values where absent. Almost all access to env goes via envutil.js.

Webpack outputs a bundled es module for target:webworker, which Wrangler then
deploys with the help of build.upload and build.upload.rules directives.