mv disk-based blocklist setup to config
svc.js is common to both node and deno runtimes, and so there is a reliance on
dynamic imports to run blocklist-setup code specific to those runtimes. Moving
that code to config.js (node-only) and config.ts (deno-only) lets us rely on
static imports, which work out-of-the-box with deno-bundle and denoctl-deploy.