auth for doh and dot
An auth token, which is hex(sha256(key + ctx)), where ctx is the domain name
of any given serverless-dns installation and key is some msg secret as
desired by the admin user.

This auth token can be checked-in into source control, put up as env var, or
registered as a secret for up to 64 chars.

The auth token is sent as the final parameter of the full blockstamp.

In DoH as part of the url path: <ver>:<base64-stamp>:<hex-auth-token>,
where auth-token can be upto 64 chars

In DoT as part of the server-name-identification (SNI) field:
<ver>-<base32-stamp>-<hex-auth-token>, where the entire SNI field can be
upto 63 chars. It is hence advisable to limit hex-auth-token at 12 to 24 chars.