dns/conns: avoid sweeping often
pooler's sweep op is quite expensive (consumes 100% cpu) when the server
is dealing with 500+ concurrent DoT / DoH requests. Basically, when these
DoH / DoT connection arrive all at once, too many outgoing connections are
created that when done wait to "checkin" it to the pool. If the pool is full
and fresh (all other connections are also recent ones), then the sweep op to
find the new checked-in connection a slot in the pool is not only unnecessary
but also expensive on the cpu as it loops over the entire length of the pool
and does nothing.