node: rmv async+await; use callback style for h2 req
Mostly to avoid scenarios where two different error handlers are used:
once for async-await (with a try-catch block), and once for callback-style
(with a listener on "error"). This is confusing especially, since any error
on the stream is sent to both the async-await block and the error callback.
If the error goes unhandled from any one of them, then node process shuts down.